Refactor email sending logic and add order emails
Moved verification email logic from AuthService to EmailService for better separation of concerns. Added methods in EmailService to send order confirmation and payment success emails.
